NT: update image build only on tag push or workflow dispatch
When using docker/metadata-action, it will create a docker image tag that matches the branch name
if there is no git tag on the commit that is pushed. In this case the "latest" docker image tag is
not advanced, which is not useful for our case.

The "latest" docker image tag is only advanced when a git tag is present on the commit, and so
for our cases only building images on tag push is the desired behavior.